requirements.txt
asgiref==3.3.4
astroid==2.6.2
Django==3.2.4
django-crispy-forms==1.12.0
django-filter==2.4.0
django-hCaptcha==0.1.0
django-js-asset==1.2.2
isort==5.9.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.3.0
psycopg2==2.9.1
pylint==2.9.1
python-dotenv==0.18.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
wrapt==1.12.1
